\name{nmaplateplot-package}
\alias{nmaplateplot-package}
\alias{nmaplateplot-package}
\docType{package}
\title{
The Plate Plot for Network Meta-Analysis Results
}
\description{
A graphical display of results from network meta-analysis (NMA). It is suitable for outcomes like odds ratio (OR), risk ratio (RR), risk difference (RD) and standardized mean difference (SMD). It also has an option to visually display and compare the surface under the cumulative ranking (SUCRA) of different treatments.
}
\author{
\strong{Maintainer}: Zhenxun Wang \email{wang6795@umn.edu}

Authors:
\itemize{
  \item Zhenxun Wang
  \item Lifeng Lin
  \item Shanshan Zhao
  \item Haitao Chu
}
}
